# Fix quadratic jQ.add fold in fragment construction.