@ehberger 2015-06-29
That's some good, clean computer science right there.  I feel like I'm
looking at a diagram in an algorithms textbook.

On Mon, Jun 29, 2015 at 2:32 PM, Jason Merrill <notifications@github.com>
wrote:

> Graph of rendering time before (orange) and after (blue) this change.
>
> [image: timing graph]
> <https://camo.githubusercontent.com/b52548e0676b8b83cf07c454668592cceb8b80ad/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f677261706865722f6578706f7274732f7365626967626d786b352e706e67>
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/desmosinc/mathquill/pull/13#issuecomment-116851989>.
>