@jwmerrill 2015-06-29
Every now and then, you get the feeling that it's possible to understand how computers work.