@jwmerrill 2015-06-29
Eli reviewed this, and gave me a verbal go-ahead to merge it.