@eluberoff 2015-06-29
confirmed

On Mon, Jun 29, 2015 at 5:54 PM, Jason Merrill <notifications@github.com>
wrote:

> Eli reviewed this, and gave me a verbal go-ahead to merge it.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/desmosinc/mathquill/pull/13#issuecomment-116874111>.
>